add plan for version 2
