+++ CLAUDE.md
@@ -1,0 +1,1 @@
+See `.github/copilot-instructions.md` for agent instructions
