+++ package.json
@@ -70,5 +70,6 @@
         "typescript": "^5.9.3",
         "typescript-eslint": "^8.45.0",
         "vitest": "^3.2.4"
-    }
+    },
+    "packageManager": "pnpm@10.26.1+sha512.664074abc367d2c9324fdc18037097ce0a8f126034160f709928e9e9f95d98714347044e5c3164d65bd5da6c59c6be362b107546292a8eecb7999196e5ce58fa"
 }
